extract helper function
