+++ src/use/crypto/formatPem.ts
@@ -1,0 +1,20 @@
+export function formatPEM(pemString: string) {
+    const PEM_STRING_LENGTH = pemString.length,
+        LINE_LENGTH = 64;
+    const wrapNeeded = PEM_STRING_LENGTH > LINE_LENGTH;
+
+    if (wrapNeeded) {
+        let formattedString = "",
+            wrapIndex = 0;
+
+        for (let i = LINE_LENGTH; i < PEM_STRING_LENGTH; i += LINE_LENGTH) {
+            formattedString += pemString.substring(wrapIndex, i) + "\r\n";
+            wrapIndex = i;
+        }
+
+        formattedString += pemString.substring(wrapIndex, PEM_STRING_LENGTH);
+        return formattedString;
+    } else {
+        return pemString;
+    }
+}
